refactor: extract LibraryHandler from MCPServerV2
Move 11 library management methods into dedicated LibraryHandler.
Move _test_library_dir to handler; update test references.
